change route to RegionDetail
